Disabled Autoplay Settings Section
This is necessary to remove something that is useless and since it also doesn’t work I have no reason to keep it shown in Settings. but this commit disables the Autoplay Settings Section. and there will not be an option for this because it would be a silly addition to CercubePlus. If anyone hates this change then fork it and remove it if you want the Autoplay Section in Settings to be enabled.